% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{rapiAsPOSIXct}
\alias{rapiAsPOSIXct}
\title{R Wrapper for \code{strptime}}
\usage{
rapiAsPOSIXct(x, tz = "")
}
\arguments{
\item{x}{Vector with R objects to be converted to \code{POSIXct}}

\item{tz}{An optional timezone string}
}
\value{
A vector with \code{POSIXct} datetime objects
}
\description{
Wrapper for C-level \code{asPOSIXct} function
}
\details{
This function provides an entry point from R to the
\code{asPOSIXct} function of the R API. It is use mainly for
testing as R itself exposes the function.
}
\author{
Dirk Eddelbuettel
}
