\name{RApiDatetime-package}
\alias{RApiDatetime-package}
\alias{RApiDatetime}
\docType{package}
\title{\packageTitle{RApiDatetime}}
\description{\packageDescription{RApiDatetime}}
\section{Package Content}{\packageIndices{RApiDatetime}}
\author{\packageAuthor{RApiDatetime}}
\section{Maintainer}{\packageMaintainer{RApiDatetime}}
\keyword{package}
