% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{rapistrptime}
\alias{rapistrptime}
\title{R Wrapper for \code{strptime}}
\usage{
rapistrptime(x, fmt, tz = "")
}
\arguments{
\item{x}{Vector with one or character elements to be parsed}

\item{fmt}{The format string, see \code{help("strptime")} for
details.}

\item{tz}{An optional timezone string}
}
\value{
A vector with \code{POSIXlt} datetime objects
}
\description{
Wrapper for C-level \code{strptime} function
}
\details{
This function provides an entry point from R to the
\code{strptime} function of the R API. It is use mainly for
testing as R itself exposes the function.
}
\author{
Dirk Eddelbuettel
}
